requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.2
Django==3.1.8
django-pwa==1.0.10
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.1.0
httplib2==0.19.0
idna==2.10
isort==5.7.0
jwcrypto==0.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.3
pep8==1.7.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
Pyrebase4==4.4.3
python-jwt==3.3.0
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
